+++ CHANGELOG.md
@@ -5,10 +5,12 @@ The format is based on [Keep a Changelog(https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning(https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Support WeasyPrint 60.0 new option
 
 ## 1.2.0 - 2023-05-11
 ### Added
-- Support WeasyPrint 59.0b1 new option
+- Support WeasyPrint 59.0b1 new options
 
 ## 1.1.1 - 2023-04-27
 ### Security