+++ CHANGELOG.md
@@ -6,9 +6,13 @@ and this project adheres to [Semantic Versioning(https://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## 1.3.0 - 2023-10-07
+### Added
+- Support WeasyPrint 60.0 new option
+
 ## 1.2.0 - 2023-05-11
 ### Added
-- Support WeasyPrint 59.0b1 new option
+- Support WeasyPrint 59.0b1 new options
 
 ## 1.1.1 - 2023-04-27
 ### Security